+++ src/treg/caller_metadata.py
@@ -5,6 +5,13 @@ import re
 from fastapi import Request
 
 
+_META_KEY_RE = re.compile(r"^[a-z0-9_]{1,32}$")
+_MAX_BUDGET_DIMS = 3        # each declared dimension = one indexed lookup per call + a row per value
+# The reserved value standing for "every value of this dimension". Safe forever because the tag
+# value charset (`_META_VALUE_RE`) excludes `*`, so no caller can send a value that collides with it.
+TAG_DEFAULT = "*"
+
+
 def _norm_client(raw: str) -> str:
     """A runtime name as a short slug. One spelling for both ends of the roster: what an incoming
     header is stored as, and what `promoted_from` must match to hide a detected pair."""
